Major Refactoring of project
This commit is the first of many that refactors the project.
It renames many functions and files as well as renaming the
project name to mod_MP instead of mod_MediaPlayer.

A lot of functinoality is still broken and lots more work
still needs to be done. Eventually this should allow for full
XML and JSON support.